Less sensitive godmode check, bugfixes for log messages, better check names

+++ src/cc/co/evenprime/bukkit/nocheat/checks/blockbreak/DirectionCheck.java
@@ -78,7 +78,7 @@ public class DirectionCheck extends BlockBreakCheck {
         switch (wildcard) {
 
         case VIOLATIONS:
-            return String.format(Locale.US, "%d", player.getData().blockbreak.directionVL);
+            return String.format(Locale.US, "%d", (int)player.getData().blockbreak.directionVL);
 
         default:
             return super.getParameter(wildcard, player);
